Added framework for log favorites
Began work on adding log 'favorites' (for lack of a better name).
These are named logs which can be used to populate new logs. The
idea being frequently used actions can be saved (such feeding
the fish in the morning and adding, say, dissolved CO2) and used
for rapidly adding log entries without much fuss.

The idea being one can simply clink on a favorite on the main
aquarium page for that action to take place.