Minor changes with aquarium view
Added a public link if the aquarium's visibility is public.
The if-checks for the template are getting kinda complicated, so I'm
pondering having a separate set of public methods and views.
Some duplication of work ends up happening but it would be much easier
to customize public views later on down the line that way while also
making it a bit easier to keep security in check, I think.