Default timezone is now America/Chicago
Default timezone is now America/Chicago for when users are not logged in
In practice, this happens through the main auth filter, so there may be
times where this might not apply. It works for now.

This was tossed in due to timezones looking off when viewing an aquarium
via the public link. In practice, it might be better to lookup the user's
timezone and use that instead.