Began implementing Aquarium Life module
Basic add/update works and some validation.
Still need to add an AquariumLifeLogs table, attach logic
And add a delete function (and make sure it's hard, not soft deletes
since deleteAt is being used as removed from tank - a bit of a hack to work better with Laravel - I may not need it).

Run on sentence. It's late.